Create README.md
Social media has become such a large influence in the modern business world. Companies/businesses need to stay up to date with customers perception of their services, the majority of which can be captured from social media. Twitter is a major social media platform. Customers are continuously providing postive and negative feedback of the services provided by companies via Twitter. Airlines are no exception to this.

The problem at hand is a form of sentiment analysis. The categories of feedback are negative, neutral and postive feeback, along with reasoning for any negtive types of feedback. We will apply Natural Learning Processes in order to provide useful insight for Airlines. We will look to clean the data, fit and tune a model which can then be used on additional data sets to provide further insight to the Airlines.

The text variable within the data is key, as it consists of the description of the specific tweet. Being able to determine key metrics and recurring areas with negative feedback, will allow the specific Airline to determine where they can improve their service.